+++ pkgs/native_assets_builder/test_data/package_reading_metadata/hook/build.dart
@@ -2,16 +2,21 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-// ignore_for_file: deprecated_member_use
-
 import 'package:native_assets_cli/native_assets_cli.dart';
 
 void main(List<String> args) async {
   await build(args, (input, _) async {
-    final someValue = input.metadatum('package_with_metadata', 'some_key');
+    final someValue = input.metadata['package_with_metadata']['some_key'];
     assert(someValue != null);
-    final someInt = input.metadatum('package_with_metadata', 'some_int');
+    final someInt = input.metadata['package_with_metadata']['some_int'];
     assert(someInt != null);
     print({'some_int': someInt, 'some_key': someValue});
+
+    // ignore: deprecated_member_use
+    final someValueOld = input.metadatum('package_with_metadata', 'some_key');
+    assert(someValueOld != null);
+    // ignore: deprecated_member_use
+    final someIntOld = input.metadatum('package_with_metadata', 'some_int');
+    assert(someIntOld != null);
   });
 }